Hephthalite work
-Additional Sogdian characters added.
-Changed Khingila into the former Alchon ruler of Gandhara/Taxila. The
history of the Hephthalites is confused and conflicted, but I /think/ he
belongs there and this Kushnavaz guy is the rightful Hephthalite Proper
leader.
-Provinces tweaked a bit around Afghanistan. Nomads don't get taxes from
regular landed holdings, /which is why some have been left in./ The
Hepthalites have an alarmingly strong military already, and they haven't
even been assigned all the provinces they ought to yet.